requirements.txt
requests~=2.32.3
pika~=1.3.2
kafka-python~=2.0.2